Avoid systemd warning about unset JAVA_OPTS environment variable
When starting the unifi service, systemd logs the warning:

    unifi.service: Referenced but unset environment variable evaluates
    to an empty string: JAVA_OPTS

Set JAVA_OPTS="" in the sysconfig file to avoid the harmless, but also
needless warning.

While here, remove a typo in the sysconfig file by slightly rewording
the comment and replace EOL with EOF as the latter more accurately
reflects the action.